backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
django-extensions==3.2.3
# psycopg2-binary==2.9.9  # Optional - only needed for PostgreSQL (using SQLite for local dev)
python-dotenv==1.0.0
Pillow>=10.2.0
django-filter==23.5
djangorestframework-simplejwt==5.3.1
celery==5.3.4
redis==5.0.1
stripe==7.8.0
python-dateutil==2.8.2
openpyxl==3.1.2
reportlab==4.0.7
qrcode==7.4.2
phonenumbers==8.13.27
# barcode - optional, install separately if needed: pip install python-barcode
# pandas and numpy - optional for analytics, install separately if needed
# pandas>=2.2.0
# numpy>=2.0.0

